Dedupe Starline{,Station}Transformer
